+++ Zeldish/Zeldish/BoundingVolume.h
@@ -7,15 +7,15 @@
 class BoundingVolume
 {
 private:
-	int m_x, m_y;
+	float m_x, m_y;
 	int m_width, m_height;
 public:
 	BoundingVolume();
 	~BoundingVolume();
 	void ShutDown();
-	int Initialize(int x, int y, int width, int height);
+	int Initialize(float x, float y, int width, int height);
 
-	int SetPosition(int x, int y);
+	int SetPosition(float x, float y);
 
 	int CheckAgainst(BoundingVolume* other);
 };
